Refactor API modules
