+++ app/api_v1/__init__.py
@@ -1,12 +1,7 @@
 from flask import Blueprint
 
 api = Blueprint('api_v1', __name__)
-from . import authentication, errors, utils
-from .errors import *
-from .index import *
-from .rate_limit import *
-from .etag import *
-from .utils import *
-from app.api_v1.resources import *
-from app.api_v1.fhir import *
-from .errors_test import *
+from app.api_v1 import resources
+from app.api_v1 import errors
+from app.api_v1 import index
+from app.api_v1 import fhir_errors
