--- app/api_v1/errors_test/__init__.py
@@ -1,2 +1,0 @@
-from .fhir_errors import *
-from .exceptions import *
