+++ app/api_v1/exceptions.py
@@ -1,4 +1,4 @@
-from ..errors import AuthenticationError
+from app.api_v1.errors import AuthenticationError
 
 class TokenExpiredError(AuthenticationError):
     pass
