--- app/api_v1/fhir/__init__.py
@@ -1,2 +1,0 @@
-from .utils import *
-from .resources import *
