--- app/api_v1/fhir/utils/__init__.py
@@ -1,3 +1,0 @@
-from .bundle_utils import *
-from .operation_outcome_utils import *
-from .search_utils import *
