+++ app/api_v1/fhir_errors.py
@@ -1,10 +1,11 @@
-from ...api_v1 import api
-from .exceptions import TokenExpiredError
+from app.api_v1 import api
+from app.api_v1.exceptions import TokenExpiredError
 from flask import jsonify, url_for
+from app.api_v1.operation_outcome_utils import create_operation_outcome
+
 
 @api.errorhandler(TokenExpiredError)
 def token_expired(e):
-    from ..fhir.utils.operation_outcome_utils import create_operation_outcome
     oo = create_operation_outcome(outcome_list=[
         {'severity': 'error', 'type': 'expired', 'location': [url_for('api_v1.new_token', _external=True)],
          'diagnostics': e.args[0], 'details': 'Authentication token expired: complete authentication for new token'}])
