+++ app/api_v1/index.py
@@ -1,7 +1,7 @@
 from flask import jsonify, url_for
-from . import api
+from app.api_v1 import api
 from app.api_v1.etag import etag
-from .authentication import token_auth
+from app.api_v1.authentication import token_auth
 
 
 @api.route('/', methods=['GET'])
