+++ app/api_v1/rate_limit.py
@@ -1,6 +1,4 @@
-import functools, os
-from flask import request, url_for, make_response
-from app.utils.general import json_serial
+import functools
 from app.api_v1.errors import *
 import time
 from app import redis
