+++ app/api_v1/resources/__init__.py
@@ -1,1 +1,1 @@
-from .User import *
+from app.api_v1.resources import User, Patient, CodeSystem, ValueSet
